requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

jsoncparser==1.1.0
websockets==12.0
